reimplement hide command on top of SQLite
The `hide` command relied on adding a special entry to the ref-log indicating that a certain commit was hidden. However, in practice, pygit2 didn't have very good APIs for adding entries to the ref-log, and it seems that attempts to add a no-op entry to the ref-log (i.e. update HEAD to HEAD, when HEAD Is a direct reference) are dropped by git/libgit2.