fix bug with traversing parent hidden commits
There was a bug where we try to delete a parent commit from the commit graph before we've finished visiting its child commits, causing a `KeyError` later. I wasn't able to add a test case to expose this, because it's a little bit non-deterministic (relies on the iteration order of sets).