reverse output ordering
The original ordering is showing the oldest commits at the bottom and the newest commits at the top. This matches how some other tools render, which makes sense in a GUI where you start at the top of a pane and scroll down.

However, in the command-line, the latest output is shown at the bottom, so it makes sense to avoid scrolling by putting the newest commits (most likely to be relevant to the user) at the bottom as well.