eventlog: replace ref-log with event log
The ref log is unfortunately insufficient for logically determining the results of user actions. In particular, it doesn't clearly tell us what the old and new commit OIDs were during a rewrite. Fortunately, this information is exposed by Git hooks, so we can record it ourselves and play it back.