graph: allow for hiding master commits
This is technically not a supported use-case, as we expect master history to be immutable and always visible. But it can happen sometimes — for example, if you commit to master locally, and then `git rebase -i origin/master`.

In those cases, the branchless invariants are broken and things might not function properly (such as commits not being visible in the smartlog that should be, or vice versa), but the situation might be a little bit better if you realize that, for example, a `master` commit is thought to be hidden by branchless. Then you can manually fix up the situation if it's causing a problem.