graph: improve handling of hidden master commits
The intent of this refactoring is to make it possible to restack hidden master commits. To do this, we need to correctly handle hidden master commits.

The key idea is to process all event-log events when building the graph, and hide them at a later stage in a more intelligent way, when we have access to the structure of the graph. Then we can surface hidden master commits in the smartlog, which will tell the user that something strange has happened and they should restack. (Surfacing hidden master commits in the smartlog also enables the `restack` command itself to work in these cases, as it requires the old hidden commit to be visible in the smartlog.)