hide: add -r flag for recursive hides
It seems to be too difficult to ascertain the user's intention to hide a single commit in a subtree. Rather than guess it and attempt to hide all children, or automatically assume that the user meant to hide the entire subtree, we just add a `--recursive` flag for the user to indicate that it's their intention.

This would be best followed up to a transactional system for `undo` to ensure that the entire hide operation can be undone in one action.