docs: remove embedded demo
The SVG demo is large and significantly slows down rendering of the webpage. It also doesn't display colors correctly. (I also tried exporting as GIF, but that produces an even larger file, although it may be better-optimized for playback in some way.) The best rendition is currently available at asciinema, so just link to that directly.