gc: keep commits live until explicitly hidden
Prior to this commit, it was possible that Git's garbage-collection mechanism would collect a commit that's still referred to in the smartlog (since there are no refs pointing to that commit or its descendants). This commit changes it so that all user commits are kept live by default, and then can be cleaned up by Git once they are hidden and `git branchless gc` has run and marked them as unreachable.