cleanup(rust): use `context` macro
The idea is to use the `context` macro for most functions which return a `Result` (or `anyhow::Result`).

`anyhow::Context` adds a `with_context` method for `Result`s, to add an informative message in the case that they return an `Error`. This is then displayed in the call-stack rendered when an error occurs. However, this means that when doing common operations, we end up duplicating the informative message for each call-site to the underlying function. This is typically unnecessary, because the message should be the same for each caller. In this case, we instead annotate the function itself with a `#[context(...)]` macro to attach the context automatically, rather than having each caller attach it manually.

For third-party libraries which haven't annotated their functions with `#[context]`, we still use `with_context` in an ad-hoc fashion.