ci(rust): fix broken `test_pre_auto_gc`
This test was failing because it was executing the hook in the current directory, rather in the temporary directory which contained the Git repository. On my local machine, it happened to be executing the previously-installed `pre-auto-gc` hook. On a CI machine, `git-branchless` was not initialized in the working directory, so it failed.

Fixed by changing the current directory to the Git repository directory before proceeding with the rest of the test.