undo: show the entire transaction's events in the UI
`git undo` was previously changed to show the effects of an entire transaction, but due to an oversight, the description of that transaction was limited to only its most recent event. This commit fixes that, and shows the entire list of transaction events.

In the future, we might need to encase the list of transaction events in some kind of UI control so that it can be scrolled without occluding the smartlog display (which is already in such a control).