refactor(tui): add macros to manage `cursive` views
There's a risk of some type confusion with Cursive's built-in view system. This commit adds a type-safe macro-based implementation on top of the views. This is also used in a future commit to implement `git-branchless checkout`.

It's possible that there's a purely type-system-based encoding of the same thing. I could only think of a higher-kinded-type-based encoding, which I think Rust doesn't support.