refactor(formatting): use `StyledString`s instead of ANSI escape codes
Before this commit, we directly interpolated ANSI escape codes into strings to produce their coloring, and then we later stripped it out if we decided that we didn't need it. Some reasons:

1) This is a bit unprincipled, because you don't know if a certain string represents a raw string which is safe to print anywhere, or a formatted string, which must only be printed to `stdout`.
2) The `cursive` APIs do not expect to be passed strings with ANSI escape codes in them. As a result, many of the components miscalculate the size of their contained text, and position things incorrectly. By using `StyledString`s, the components are able to calculate the correct logical size for the provided strings.
3) When rendering the smartlog, we recursively render child subtrees, then append a border to those lines. For deeply-nested subtrees, this results in potentially O(n^2) behavior. This commit still has the O(n^2) behavior, but it might be a little bit less expensive in practice since we're combining `Vec`s instead of `String`s. A proper solution might store a `Vec` of `StyledString`s in reverse order, and only reorient them at the end.

This commit depends on https://github.com/gyscos/cursive/pull/580. For this commit, we use a `git` dependency in `Cargo.toml` in order to ensure we have the right code. However, packages with `git` dependencies can't be uploaded to https://crates.io, so for the next `git-branchless` release, we will either have to fork upstream `cursive` into our own crate, or request that they publish a new version with the necessary change.