cleanup: remove most `&mut impl Write`s
Most of these were in place to make testing easier, but in practice, we run integration tests and collect the stdout/stderr directly. These parameters just add bloat to the interfaces without adding value.

Additionally, it becomes hard to work with the provided `Write` instances. We can't check to see if they're a TTY to decide whether or not to write colors, and the current implementation of `run_git` waits until all of the output has been provided and then returns it in a `String`, rather than stream it to the user.

It would probably be best to make our own input/output type and wrap all IO using that, rather than use `Write` as a trait, which is insufficient for our needs.