Invoke hooks using sh/bash.exe explicitly
To run correctly on Windows, all invocations of hooks need to be
executed as an argument to bash.  This changes the code to build
commands using get_bash() with the `-c` flag.

When discovering which bash to use, we are careful on Windows to pick
"git-bash" aka the bash.exe that comes with git.  This is neccesary as
git-bash uses different path translations than regular bash.exe, and
using the wrong one would result in git.exe not being able to discover
the git-branchless.exe binary correctly (when invoked in hook).

If the target_os isn't windows, `sh` is for executing hooks instead.