Set a default_path for Windows git.exe
In tests, when the "system git" is desired, PATH_TO_GIT is ignored and a
hardcoded path to the system git is used (typically /usr/bin/git).

"system git" is only used in two places, and it is unclear how important
those locations be handled by "system git" rather than "git under test",
so just do the conservative thing and add a path to the default location
of git.exe on Windows.