feat(rewrite): don't query repository to generate label names
As part of generating rebase plans, we generate temporary label names, which are used to refer to a commit which hasn't been made yet (and for which we therefore don't know the commit hash).

This is the relevant part of `git-rebase(1)`:

> The label command associates a label with the current HEAD when that command is executed. These labels are created as worktree-local refs (refs/rewritten/<label>) that will be deleted when the rebase finishes. That way, rebase operations in multiple worktrees linked to the same repository do not interfere with one another. If the label command fails, it is rescheduled immediately, with a helpful message how to proceed.

Before this commit, we would preemptively check to make sure that there was no ref under `refs/rewritten` in the appropriate name. However, given the above, it seems that we can rely on Git to delete everything from `refs/rewritten`, and therefore start with an empty `refs/rewritten` directory when we start the rebase operation. Thus, we can avoid making several I/O queries to Git.