fix(rewrite): treat branch names more carefully
The basic problem is that reference names provided by the user may not be unambiguous. The reference `abc` might refer to a reference under `refs/heads/` or `refs/remotes/` or at the top level of the `.git` directory (or even an object hash).

The user provides a main branch name without the necessary prefix. We want to support both local and remote main branches, so we have to figure out which kind of branch it is. Due to sloppy handling of branch names, given the main branch name `master`, sometimes we accidentally created a top-level reference called `master`, instead of the appropriate branch `refs/heads/master`.

This commit improves the discipline around branch name handling. The rule is: we should never use shorthand branch/reference names in the internal workings of the system; these should be generated at the time we wish to display the shorthand branch name to the user. Various parts of the code have been reworked to support this. It's worth noting that reference names are always `OsString`s, while names which are to be displayed are `String`s, so hopefully the type system will help us avoid mixing these up.