feat(rewrite): calculate patch IDs in parallel
This step seems to take a lot of times when we're trying to rebase an old branch onto the latest main branch, because we have to detect duplicate commits across a large range.

In the long run, it would probably be best to filter out the majority of the upstream commits by checking to see if they touch any of the same files as a local commit.