refactor(tui): rename `Output` to `Effects`
The intention is for the `Effects` type to handle all sorts of effects. It started as just output, but it also handles progress indication, and might expand to more things in the future.

We currently don't need any fancy effect features, but we could use https://github.com/pandaman64/effective-rust/ in the future if we do. In particular, it's not currently obvious what particular effects a function consumes. There are several functions which report progress but do no output, which would be nice to know from the type signature. We also don't need any of the continuation features at the moment.