feat(rewrite): calculate patch IDs in parallel
When initiating a rebase, it seems that Git calculates patch IDs in order to detect duplicate commits and decide not to apply them again. Calculating the patch IDs for all of the upstream commit can be quite expensive, so parallelize it to improve the performance.

That being said, it seems that Git is significantly faster at detecting the duplicate commits, so there is probably some misunderstanding in the above about how Git detects them.