fix(undo): report unsupported use-case for empty events
Closes #57. Some operations, such as bisecting and stashing, use references in an unconventional way. These are reported as if there is no change to the content of the reference, which might happen e.g. if instead the reflog for that reference was changed, or if it was touched but left unchanged.

These uses for references don't really affect the commit graph. Ideally, we would handle them entirely differently. They should have an entirely separate history, and it should be possible to rewind them separately from the commit graph's history. For example, you should be able to unmark a certain commit during a bisect operation without affecting any commits you may have made since starting the bisect.

The UI for this would need to be considered thoughtfully. Since this isn't too common of a use-case for now, I'm updating this message to just warn that it may be an unsupported use-case, and deferring it the alternative UI for later.