docs: downgrade `git-branchless` to alpha
`git-branchless` was initially described as being in beta. However, it turns out that an implementation of commit evolution was not sufficient for me to bring `git-branchless` suitably up to parity with Mercurial, so the scope has broadened. I also think that some workflows could be improved beyond what either Git or Mercurial currently offer.