fix(tui): fix race condition when printing subprocess output quickly
Under the following circumstances:

- It's been less than 250ms since program startup.
- An `OperationType` has been started.
- Output is then printed via `effects.get_output_stream()`/`effects.get_error_stream()`

The output would be suppressed. This is because the `MultiProgress` would still be hidden, and it would suppress any output sent to it.

However, other kinds of output which are printed before an operation has started would still be displayed.

This commit fixes it by adding a `bool` which indicates whether the `MultiProgress` is currently hidden.