fix(tui): don't overwrite output from subcommands after 250ms
Previously, output from subcommands could be suppressed entirely if the command completed before 250ms. This is because the updater thread would not `tick` the `MultiProgress`, so calls to `ProgressBar::println` with the `MultiProgress` as the `draw_target` would do nothing.

To fix this, I added an `is_visible` flag to each `OperationState`. Once the 250ms timeout elapsed, the updater thread would set the `is_visible` flag for each operation.

However, this only set the flag for operations that were underway when the 250ms timeout elapsed. If other operations start, or if an operation finishes and is removed from the `operation_states` map and is added again later, then the `is_visible` flag would be set to `false` again, even though it's now safe to print above the `MultiProgress`. As a result, output would go directly to `stdout`. The next `tick` of the progress bar would immediately overwrite the most recent line, which means that the user would potentially miss most of the lines of output. Additionally, a copy of the progress bar at a previous point in time would be snapshotted and left in the output, which looks ugly.