feat(logging): add `RUST_PROFILE` environment variable to trigger profiling
The `RUST_PROFILE` variable can be set to emit a Chrome logging trace. The result can be opened in `chrome://tracing` in Google Chrome, or in another viewer.

If the `RUST_PROFILE` variable is set to `1` or `true`, then it automatically chooses a filename for the profile. Otherwise, it uses the value of the `RUST_PROFILE` variable as the name of the file to write to.

Initial tracing indicates that the biggest time-consumers for a simple `cargo run smartlog` are 1) walking from commits and 2) loading all events from the event log.