fix(rewrite): only detect upstream commits with the exact same changed paths
Previously, we calculated the patch ID for any commit which had at least one path in common. The idea was that it would be safer in the case that the logic was incorrect (such as around handling renames).

The original motivation for this commit is to handle the case of a big upstream commit, in which case 1) it's more likely to trigger a false positive here and induce a diff to be calculated, and 2) the calculated diff will naturally take more time.

But upon further reflection, it also just... doesn't make sense to do it the old way, because the patch IDs will never be the same if the changed paths are not the same.