feat: add -C option to set working directory
`git` has a `-C` option to set the working directory before proceeding with the rest of the operation. This commit adds the same `-C` option to `git-branchless`.

The original motivation for this is to run the binary using `cargo run`, but on a different repository. But it might be useful for other things as well.