fix(rewrite): various optimizations for calculating changed paths
The most important optimization by far is removing `#[instrument]` from the hot recursively-called function.

I also added the following as optimization attempts, and we might as well keep them:

- Loading `Tree` objects lazily.
- Avoiding loading a given `TreeEntry` by reading the tree contents from disk again; instead, keep a reference directly to it.
- Reducing memory allocations by allocating path components in a `Vec`, rather than appending directly to a `PathBuf` (which appears to be backed by an `OsString`, so appending components repeatedly would require many copies).

Benchmarks indicate only around a 2x performance speedup on a problematic commit:

```
diff/Repo::get_patch_for_commit
                        time:   [1.0360 ms 1.0512 ms 1.0668 ms]
                        change: [-51.133% -50.086% -49.187%] (p = 0.00 < 0.05)
                        Performance has improved.
Repo::get_paths_touched_by_commit
                        time:   [98.755 ms 104.32 ms 118.59 ms]
                        change: [-69.251% -58.564% -43.885%] (p = 0.00 < 0.05)
                        Performance has improved.
```

But for whatever reason, it seems to actually produce a speedup on the order of 100x in practice when part of a `git move` operation (>100s -> <1s). I'm not sure why this is. It might be that the effect is only really noticeable when the appropriate files on disk haven't been paged into memory in some way. This would also explain why Criterion reports a much higher estimated time to carry out the sampling than it actually takes: the first operation would take a long time, but the rest would be much faster.