refactor(repo): create `Repo::get_references_snapshot`
This function subsumes the previous three functions, which got the `HEAD` OID, the main branch OID, and the OIDs for any branches. The intention is to make sure that this function also updates the `Dag`, so that any values derived from the return result are known to be safe to use in `Dag` queries. It's also just more convenient to work with, rather than the three separate local variables/parameters.