refactor(dag): remove inner `RefCell`
The inner `RefCell` was because we may have had to borrow it mutably when it was satisfying the migratory `MergeBaseDb` trait. Since we've gotten rid of the trait, we can accurately track mutability now.

The approach before this commit was to try to update the DAG whenever we saw an OID that it didn't know about. The idea going forward is to do updates to the DAG as a single step, immediately after getting the set of references for the repository, and then using it immutably from there.