graph: fix performance issue when calculating visible heads
When calculating anomalous main branch commits, we query for all of the main branch commits and intersect it with obsolete commits. These commits are to be added to the smartlog to surface to the user, as they're probably anomalous.

The number of main branch commits will typically far exceed the number of obsolete commits. Thus, it's much faster to iterate over the set of obsolete commits and do `contains` checks on the set of main branch commits (which each can be completed fairly quickly), rather than iterate over the set of main branch commit.

This commit reduces the smartlog rendering time from >10s (a regression from the pre-DAG implementation!) to <1s (an improvement from ~3s from the pre-DAG implementation).