fix(smartlog): don't show active public commits unnecessarily
This is a subtle bug which arose due to the migration to the Eden DAG.

The `DagAlgorithm::heads` function returns the *heads* of the provided set, which is the set of elements which do not have a child element also in the same set. We apply the `heads` function to the set of all observed commits to get the heads to render. Under all of the existing test cases, it happens that every public commit is also an observed commit, because the `post-commit` hook would observe them, so it always happened that the head of the set of public commits was the main branch commit.

In an actual repository, not every public commit will have been observed by the `post-commit` hook, so it's possible to have "holes" separating the sequences of observed public heads. Since the `heads` function finds vertexes which do not have a child vertex *in the same set*, it would find and return isolated public commits in the result.

For example, if you clone a repository, and then run `git unhide` on an old public commit (specifically one that isn't already a parent of another visible commit), then it would appear in the smartlog. The pre-DAG version of the code would imperatively find and remove every commit which should be hidden, so it didn't matter if we returned extra heads for correctness. But the post-DAG version of the code does declarative set operations, so the queries have to return precise results.