refactor(graph): rename `CommitGraph` -> `SmartlogGraph`
We'll use the `Dag` type as the actual commit graph from now on. The smartlog commit graph had the problem that it didn't make it easy to query for children commits. This caused some bugs; for example, the `git next` command would not be able to find the next commit to go to if it didn't appear in the smartlog, and would instead leave you on the current commit.