feat(init): disable `man-pages` feature by default
Generating man-pages and mucking with aliases is pretty complex architecturally. I don't think it's worth keeping this code in the long-term. Instead, it's probably worth adjusting the alias definition so that it doesn't produce a warning like "no man-pages found", which can be done by changing it from a regular alias to a shell alias (starting with `!`).