fix(git): use builtins to test hook working directory
As per https://github.com/arxanas/git-branchless/discussions/120, this test uses the `cat` command, which is not a shell built-in, and has to be looked up on `PATH`. When we run the Git hook, we've removed most entries from `PATH`, so we wouldn't ordinarily be able to find it. This test happens to pass on most systems, because `git` happens to be installed in the same directory as `cat`, and we specifically add the directory containing `git` to `PATH`.

Under NixOS, the `git` executable probably isn't in the same directory, so this test fails. The fix is to avoid using a shell built-in to test this behavior.

Co-authored-by: Waleed Khan <me@waleedkhan.name>
Co-authored-by: Henri Menke <henri@henrimenke.de>