feat(smartlog): use upstream instead of local main branch, if present
A novice `git-branchless` user would probably commit directly to `master` and then not see their new commit in the smartlog, because `master` is the local main branch. We can get around this by using the upstream branch for `master` instead, since it will show the difference between the upstream and local versions of the main branch.