refactor: use `mod.rs`
Rust supports two methods of creating a multi-file module. You can either create a directory named `foo` with a file at `foo/mod.rs`, or you can create a directory named `foo` and put a file named `foo.rs` alongside it. So far, git-branchless has been using the second form, because it's supposed to be more modern. But in practice, I've found that this makes it hard to figure out if a module is made up of multiple files or not and to quickly see all of the relevant files in my editor's sidebar. So I've opted to use `mod.rs` instead, since all of the related files will now live together.