refactor: move `effects.rs` to `core`
The "effects" refer to any sort of side-effect. It's currently used for progress reporting and managing output streams, but it might be used for more things in the future, such as spawning thread/process pools. Lots of things under `core`, which indicate that it doesn't belong in the `tui` module; the `tui` (and `commands`) modules should be strictly dependent on `core`, and not the other way around.