refactor(dag): move to `core`
The DAG structure is fundamentally linked with the event log, since the event log knows about the status of visible/obsolete commits, etc.

Besides that, the `eden_dag` dependency is a little bit heavyweight, so in the event that we want to split the `git` directory into its own crate, this will make that easier.