fix(testing): fix PTY tests when running pre-commit hooks
When running `git commit` with a pre-commit hook, it sets the environment variable `GIT_INDEX_FILE`, which causes subprocess Git invocations to use an index file for the wrong repository. This causes the tests to fail.

In `portable-pty==0.5.0`, it wasn't possible to clear the inherited environment when spawning the subprocess. In `portable-pty==0.6.0`, after https://github.com/wez/wezterm/pull/1275, this is now possible. The fix is to clear the environment and set the expected base environment for spawning the subprocess.