fix(effects): fix for often-updated nested operations
It seems that nested operations which were updated frequently weren't updated correctly. The order in which we called functions like `set_message` and `set_prefix` was wrong, and the state of the progress bars would get reset, so that they appeared only as blank indicators. I think it used to appear to work before when these operations got a chance to render again without resetting the previous state, but when operations are refreshed frequently, they may not get this chance.