feat(record): remove `parent`
This was originally to support a three-way merge, but right now, it doesn't make sense, because the `before` field is assumed to be deletions and the `after` field is assumed to be additions. This probably isn't the case for most three-way merges (most likely, both sides have both additions and deletions). We can return to this feature later.