refactor(record): use `String` instead of `&'a str` in `RecordState`
As discussed with @martinvonz, this is probably more useful in practice, and the majority of the time is going to be spent by the user selecting changes, rather than in copying strings.

In the future, if we want to be able to borrow the underlying string content, we could either refactor this to use a `Cow<str>' or parameterize this on the line type altogether.