refactor(amend): use statuses indiscriminately in `git amend`
When @bcongdon originally implemented this, I recall that I told them to add a fast path where we first checked the index for any staged changes and used those before we did a scan of the entire working copy.

That's a sensible idea, but now we want to take a snapshot of the working copy before we carry out an amend so that we can undo the operation later. You could snapshot just the staged changes, but those probably aren't meaningful by themselves without the associated unstaged changes. (Or are they?)

So this commit changes it to take a working copy snapshot indiscriminately, which means we might as well unify the two code paths here.