refactor(reword): `build_messages` -> `prepare_messages`
As we move to support bulk editting, we'll soon be doing more than "building"
messages to be used from predefined inputs. "Prepare" is more descriptive of
what we'll be doing since we'll only be prepping the messages, but the user
will be doing most of the "real" work in their editor.

Also, there is no need for `PrepareMessagesResult` to be public, and its fields
were reordered into alphabetical order.