refactor(core): return `ExitCode`
Usually, we return `Result<isize>`. Since `Result` is marked as `#[must_use]`, you can't ignore the return value without producing a warning.

However, it doesn't stop situations like this:

```rust
something_which_returns_exit_code()?;
```

because the `?` operator consumes the `Result` and leaves an `isize` unconsumed. This means that we might not propagate the exit code in all cases.

To deal with this, this commit creates an `ExitCode` wrapper type which is also marked with `#[must_use]`, to ensure that the inner `Result` value is also consumed.