fix(sync): explicitly close stdin for child processes
In some circumstances, `git fetch` will hang because it's waiting on credentials, but this isn't surfaced to the user. To prevent that from happening, close the child process's stdin unless we have something to write to it.

Actually, upon testing this commit, it doesn't seem to work, but it also probably doesn't hurt to check in. We would want to suppress the "askpass" mechanism somehow (see the `GIT_ASKPASS` environment variable).