fix(sync): fix performance regression from #418
Unfortunately, adding garbage collection references for every invisible head is a significant performant regression. To fix this, we instead only mark commits as reachable if they're *active* (would appear in the smartlog). This does mean that commits which were available in the DAG but not used by the smartlog might get garbage collected, but it seems like an acceptable trade-off.