feat(rebase): Support rebasing ranges of commits, not just entire subtrees
Individual commits are included too, as they are just special "ranges" where
the source and head commits happen to be the same.

Note that the tests added to the rebase planner are very similar to the tests
used for `git branchless move`. The tests for the rebase planner are covering
situations not currently possible through `git branchless move`, though, and
I find the `smartlog` output very intuitive for such tests.