test(record): include leading whitespace in TUI tests
I think we used to trim whitespace because `insta` didn't work well with it, but it's gotten better now, so we can add the leading whitespace. Trailing whitespace gets stripped by my editor and is difficult to verify using only one's eyes since it's invisible, so we continue trimming that.