feat(reword): support rewording merge commits
We don't support rebasing merge commits in-memory at present because we don't have a method for resolving conflicts that might arise as part of the attempt to create a new merge commit. Consequently, `git reword` also doesn't work on merge commits, since it uses the rebase machinery. Since `git reword` is guaranteed not to change the contents (or topology) of any of the reworded commits, we can make an exception for this case, and permit rebasing merge commits in memory when a replacement commit is specified, because that guarantees that we won't have to resolve merge conflicts.