feat(test): set `indicatif` package's `improved_unicode` feature
I noticed some poor terminal rendering for `git test`, and I'm hypothesizing that the presence of either ANSI control codes or Unicode characters is causing it. To attempt to fix the latter, enable the feature in `indicatif` that will let it measure Unicode widths better and hopefully resolve the issue.