feat(smartlog): remove "anomalous main branch commits"
The idea of anomalous main branch commits, which is to say public commits that have been marked as obsolete for some reason, has not been very useful in practice. It also complicates describing the behavior of what commits appear in the smartlog. This commit removes it to simplify the implementation.