refactor(test): don't serialize `TestStatus` to disk
Instead, just store the exit code directly. Most of the `TestStatus`es don't apply in conditions that we write the test status to disk, so it doesn't really make sense to serialize it. Additionally, the exit code should probably be separate from the pass/fail status of a test. For bisection, you might want to invert the success condition, in which case non-zero would indicate a pass.

I also included `command` in the serialization for completeness.