refactor(dag): unify definition of commit visibility
Previously, there were a lot of different, inconsistent ways that we checked to see if a commit was "visible". Most often, we would take the set of observed commits and subtract the set of obsolete commits. However, it's possible for a commit to be obsolete and visible, when it has a non-obsolete descendant commit.

This commit centralizes the definition of commit visibility and changes the APIs to make it easier to do the right thing.