feat(smartlog): Support for rendering a sparse smartlog
This should only happen when the user has requested to render a specific
revset, as opposed to calling `smartlog` without any arguments, or when
any of the other commands trigger a call to `smartlog`.

Note that HEAD and the head of the main branch are always displayed,
regardless of the revset provided by the user.