feat(smartlog): render merge commits without duplication of subtrees
Notably, we don't follow the `git log --graph` strategy of drawing lines all the way from parent to child. These tend to cross and confuse the reader. Instead, we just print one commit in the direction of the merge and let the user resume reading at the other end of the merge (presumably by searching the rest of the smartlog output). In practice, this seems significantly easier to read, at least for me.

This also accomodates sparse smartlogs, where both ends of the merge wouldn't necessarily even be present in the smartlog. In that case, there simply won't be another end of the merge to resume from. `git log --graph` fundamentally can't do that.