refactor(rewrite): create `RebaseCommand::Replace`
We already have a mechanism for commit replacement for merge commits specifically, to handle cases where we couldn't compute the merge from scratch (and where the result of the merge should be the same anyways). This commit generalizes the mechanism to work for all kinds of commits, so that we can replace arbitrary commits with different trees without reapplying patches.

I think this also fixes a bug in merge commits where the parent order was not preserved when conducting a new merge.